+++ views/static/attraction.js
@@ -98,13 +98,15 @@ const init = async () => {
     const signInResult = await signInPromise;
     
     if (signInResult.data) {
-        showBlock(memberBtn);
         hideBlock(gateBtn);
+        showBlock(memberBtn);
+        showBlock(signOutBtn);
     }
 
     else{
         showBlock(gateBtn);
         hideBlock(memberBtn);
+        hideBlock(signOutBtn);
     }
 
     await getData(url());
