+++ views/static/index.js
@@ -167,11 +167,13 @@ const init = async () => {
     if (signInResult.data) {
         hideBlock(gateBtn);
         showBlock(memberBtn);
+        showBlock(signOutBtn);
     }
 
     else{
         showBlock(gateBtn);
         hideBlock(memberBtn);
+        hideBlock(signOutBtn);
     }
 
     const result = await getData(url(0, ""));
