+++ views/templates/index.html
@@ -64,17 +64,17 @@
         <div class="navbar">
             <div class="navbar__wrapper">
                 <a href="/"><h1 class="navbar__title">台北一日遊</h1></a>
+                <button class="navbar__menu-btn" id="js-navbar__menu-btn">
+                    <i class="bx bx-menu"></i>
+                </button>
 
-                <div class="navbar__btn-wrapper va">
-                    <button class="navbar__btn" id="js-navbar__booking-btn">預訂行程 </button>
-                    <button class="navbar__btn" id="js-navbar__gate-btn">登入/註冊</button>
+                <div class="navbar__btn-wrapper va" id="js-navbar__menu">
                     <button class="navbar__btn hidden" id="js-navbar__member-btn">會員中心</button>
-
-                    <div class="navbar__menu">
-                        <button class="navbar__btn" id="js-navbar__sign-out-btn">登出系統</button>
-                    </div>
-
-                    <i class="navbar__icon" id="js-navbar__icon"><span class="navbar__number" id="js-navbar__number"></span></i>
+                    <button class="navbar__btn" id="js-navbar__booking-btn">預訂行程
+                        <i class="navbar__icon" id="js-navbar__icon"><span class="navbar__number" id="js-navbar__number"></span></i>
+                    </button>
+                    <button class="navbar__btn" id="js-navbar__gate-btn">登入/註冊</button>
+                    <button class="navbar__btn hidden" id="js-navbar__sign-out-btn">登出系統</button>
                 </div>
             </div>
         </div>
